# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AuthUser": {
        "description": "All that other modules ever see of a user: no password hash, no ORM object.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ],
        "title": "AuthUser",
        "type": "object"
      },
      "CandidateOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "likes": {
            "title": "Likes",
            "type": "integer"
          },
          "match_score": {
            "title": "Match Score",
            "type": "number"
          },
          "my_vote": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Vote"
          },
          "place_id": {
            "title": "Place Id",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "place_id",
          "rank",
          "match_score",
          "tags",
          "likes",
          "my_vote"
        ],
        "title": "CandidateOut",
        "type": "object"
      },
      "CreateRoomIn": {
        "properties": {
          "display_name": {
            "examples": [
              "Hà"
            ],
            "maxLength": 30,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "CreateRoomIn",
        "type": "object"
      },
      "GeoPoint": {
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "title": "GeoPoint",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JoinRoomIn": {
        "properties": {
          "display_name": {
            "examples": [
              "Hà"
            ],
            "maxLength": 30,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "JoinRoomIn",
        "type": "object"
      },
      "JoinedOut": {
        "description": "Returned exactly once. The client stores the token (per room code) and sends it back in\nthe `X-Participant-Token` header on every following request.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "is_host": {
            "title": "Is Host",
            "type": "boolean"
          },
          "participant_id": {
            "format": "uuid",
            "title": "Participant Id",
            "type": "string"
          },
          "participant_token": {
            "title": "Participant Token",
            "type": "string"
          }
        },
        "required": [
          "code",
          "participant_id",
          "participant_token",
          "is_host"
        ],
        "title": "JoinedOut",
        "type": "object"
      },
      "LoginIn": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "MatchCriteria": {
        "description": "Search criteria shared by AI Match, Group Session and Food Journey.",
        "properties": {
          "categories": {
            "description": "e.g. ['cafe', 'main_meal']",
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "mood": {
            "examples": [
              "quán yên tĩnh để học bài"
            ],
            "maxLength": 300,
            "minLength": 1,
            "title": "Mood",
            "type": "string"
          },
          "origin": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPoint"
              },
              {
                "type": "null"
              }
            ]
          },
          "price_max": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "VND per person",
            "title": "Price Max"
          },
          "price_min": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "VND per person",
            "title": "Price Min"
          },
          "radius_m": {
            "default": 2000,
            "maximum": 10000.0,
            "minimum": 200.0,
            "title": "Radius M",
            "type": "integer"
          }
        },
        "required": [
          "mood"
        ],
        "title": "MatchCriteria",
        "type": "object"
      },
      "ParticipantOut": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_host": {
            "title": "Is Host",
            "type": "boolean"
          },
          "votes_cast": {
            "title": "Votes Cast",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display_name",
          "is_host",
          "votes_cast"
        ],
        "title": "ParticipantOut",
        "type": "object"
      },
      "RegisterIn": {
        "properties": {
          "password": {
            "maxLength": 72,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "examples": [
              "ha_nguyen"
            ],
            "pattern": "^[A-Za-z0-9_.]{3,30}$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterIn",
        "type": "object"
      },
      "RoomStateOut": {
        "properties": {
          "candidates": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "title": "Candidates",
            "type": "array"
          },
          "changed": {
            "default": true,
            "title": "Changed",
            "type": "boolean"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "majority_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Majority Threshold"
          },
          "me": {
            "format": "uuid",
            "title": "Me",
            "type": "string"
          },
          "participants": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ParticipantOut"
            },
            "title": "Participants",
            "type": "array"
          },
          "poll_interval_ms": {
            "title": "Poll Interval Ms",
            "type": "integer"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          },
          "status": {
            "enum": [
              "lobby",
              "voting",
              "decided",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "winner_place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Place Id"
          }
        },
        "required": [
          "code",
          "status",
          "version",
          "me",
          "expires_at",
          "server_time",
          "poll_interval_ms"
        ],
        "title": "RoomStateOut",
        "type": "object"
      },
      "StartIn": {
        "properties": {
          "criteria": {
            "$ref": "#/components/schemas/MatchCriteria"
          }
        },
        "required": [
          "criteria"
        ],
        "title": "StartIn",
        "type": "object"
      },
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "created_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VoteIn": {
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "liked": {
            "title": "Liked",
            "type": "boolean"
          }
        },
        "required": [
          "candidate_id",
          "liked"
        ],
        "title": "VoteIn",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FoodMaps API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "description": "Check credentials and set the session cookie.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "M1 auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "Clear the session cookie.",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Logout",
        "tags": [
          "M1 auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "description": "Called by the frontend on startup to tell guests from logged-in users (401 = guest).",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Me",
        "tags": [
          "M1 auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "Create an account and log in immediately (sets the session cookie).",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "M1 auth"
        ]
      }
    },
    "/api/groups": {
      "post": {
        "description": "Create a room right away (criteria come later). The creator becomes the host.",
        "operationId": "create_room_api_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Room",
        "tags": [
          "M5 group session"
        ]
      }
    },
    "/api/groups/{code}": {
      "get": {
        "description": "Polling endpoint. If since_version == room.version, returns changed=False with an empty body.",
        "operationId": "get_state_api_groups__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Version"
            }
          },
          {
            "in": "header",
            "name": "X-Participant-Token",
            "required": true,
            "schema": {
              "title": "X-Participant-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get State",
        "tags": [
          "M5 group session"
        ]
      }
    },
    "/api/groups/{code}/finalize": {
      "post": {
        "description": "Host ends voting early: the most-liked place wins (decided_by=\"host\").",
        "operationId": "finalize_api_groups__code__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Participant-Token",
            "required": true,
            "schema": {
              "title": "X-Participant-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize",
        "tags": [
          "M5 group session"
        ]
      }
    },
    "/api/groups/{code}/join": {
      "post": {
        "description": "Join the lobby. 409 if the room already started, is full or the name is taken; 410 if expired.",
        "operationId": "join_room_api_groups__code__join_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRoomIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "410": {
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Join Room",
        "tags": [
          "M5 group session"
        ]
      }
    },
    "/api/groups/{code}/start": {
      "post": {
        "description": "Host only: fetch candidates from the CandidateProvider, store them, move to voting.",
        "operationId": "start_api_groups__code__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Participant-Token",
            "required": true,
            "schema": {
              "title": "X-Participant-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start",
        "tags": [
          "M5 group session"
        ]
      }
    },
    "/api/groups/{code}/votes": {
      "post": {
        "description": "Upsert a vote, then apply rules.decide; the room may be decided in the same transaction.",
        "operationId": "vote_api_groups__code__votes_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Participant-Token",
            "required": true,
            "schema": {
              "title": "X-Participant-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vote",
        "tags": [
          "M5 group session"
        ]
      }
    },
    "/api/health": {
      "get": {
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "core"
        ]
      }
    }
  }
}